requirements.txt
django==1.6.2
django-haystack==2.1.0
django-extensions==1.3.3
django-registration==1.0
psycopg2==2.5.2
requests==2.2.1
pyelasticsearch==0.6.1
factory_boy==2.3.1
geopy==0.9.8
django-leaflet==0.11.1
django-geojson==2.3.0
django-extra-views==0.6.4
django_compressor==1.3
